chatbot-backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.25.0
pydantic>=2.6.4
python-dotenv>=1.0.1
requests>=2.31.0
openai>=1.12.0
python-multipart>=0.0.9